requirements.txt
langchain
python-dotenv
langchain-mistralai
langchain-community
pypdf
openai
tiktoken
faiss-cpu